+++ client_response.go
@@ -7,7 +7,7 @@ import (
 
 // ClientResponse represents the response object of a RPC call.
 type ClientResponse struct {
-	StatusCode  int16
+	StatusCode  int32
 	Headers     Headers
 	Content     []byte
 	ContentType string
